+++ features/settings/ui/src/commonMain/kotlin/app/campfire/ui/settings/panes/DownloadsPane.kt
@@ -315,8 +315,7 @@ private fun ItemDownloadImage(
             }
 
             drawContent()
-          }
-        ,
+          },
         contentAlignment = Alignment.Center,
       ) {
         Icon(